requirements.txt
pyrogram==1.0.1
feedparser
apscheduler
psycopg2-binary
sqlalchemy==1.3.23
python-dotenv